Add files via upload
Write a program in which linked lists are used to maintain the data structure for a car
rental company. The program allows several types of transactions to be applied to the data
structure in order to keep the lists up-to-date.
In the program three linked lists are maintained for cars:
• available for rent,
• rented,
• in repair.
The cars on the available list are ordered by mileage, with the car having the least miles at the
front of the list. The cars on the rented list are ordered by the expected return dates, with the
first car on this list having the earliest expected return date.
The program prompts the user for a transaction code (integer) as follows:
(a) add a new car to the available-for-rent list,
(b) add a returned car to the available-for-rent list,
(c) add a returned car to the repair list,
(d) transfer a car from the repair list to the available-for-rent list,
(e) rent the first available car,
(f) print all the lists,
(g) quit.